tradingbot/requirements.txt
ccxt
pandas
numpy
torch
ta-lib
websockets
python-dotenv
python-telegram-bot